🩹 chore: 파일 위치 변경 (Review.tsx, Header.tsx, ArtistCard.tsx)
